requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# OpenAI SDK
openai==1.10.0

# Queue & Background Jobs
redis==5.0.1
rq==1.16.1

# Database (for logging)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
